requirements.txt
Babel==2.6.0
Django==2.2
django-crispy-forms==1.7.2
django-phonenumber-field==2.4.0
phonenumbers==8.10.11
Pillow==6.0.0
psycopg2==2.8.2
pytz==2019.1
sqlparse==0.3.0
